Rfactoring geometry.
1) /gamegeom
-> Parallelepiped.
-> PrallelogramHorizontal.
-> ParallelogramVertical.

2) Refactoring imports, calls and etc in other Classes.
